[IIO] bugfixes related to errors from yocto run
1. set base src as async, resolves the error of hanging
  without setting base_src as async, _start_complete gets called twice
2. initializing log_file to NULL
3. NULL check for variable before using it for remove/rmdir
4. remove _query function, was incomplete, and complete function
  was same as parent class function

resolve #1353

+++ tests/nnstreamer_source/unittest_src_iio.cpp
@@ -215,6 +215,8 @@ make_iio_dev_structure (int num)
   iio_dev->dev_device_dir =
       g_build_filename (iio_dev->dev_dir, device_folder_name, NULL);
 
+  iio_dev->log_file = NULL;
+
   PREV_IIO_DEV_DIR = IIO_DEV_DIR;
   IIO_DEV_DIR = g_strdup (iio_dev->dev_dir);
 
@@ -697,7 +699,7 @@ static gint
 safe_remove (const char *filename)
 {
   /** cover for both regular file as well as pipes */
-  if (g_file_test (filename, G_FILE_TEST_EXISTS)
+  if (filename && g_file_test (filename, G_FILE_TEST_EXISTS)
       && !g_file_test (filename, G_FILE_TEST_IS_DIR)) {
     return remove (filename);
   }
@@ -714,7 +716,7 @@ safe_remove (const char *filename)
 static gint
 safe_rmdir (const char *dirname)
 {
-  if (g_file_test (dirname, G_FILE_TEST_IS_DIR)) {
+  if (dirname && g_file_test (dirname, G_FILE_TEST_IS_DIR)) {
     return rmdir (dirname);
   }
   return 0;
